+++ src/perf_trace.rs
@@ -159,6 +159,27 @@ pub mod inner {
         }};
     }
 
+    #[macro_export]
+    macro_rules! add_single_trace {
+        ($title:expr) => {{
+            use $crate::perf_trace::{
+                compute_indent, compute_indent_whitespace, format, AtomicUsize, Colorize, Instant,
+                Ordering, ToString, NUM_INDENT, PAD_CHAR,
+            };
+
+            let start_msg = "Trace".blue().bold();
+            let title = $title();
+            let start_msg = format!("{}:   {}", start_msg, title);
+
+            let indent_amount = 2 * NUM_INDENT.fetch_add(0, Ordering::Relaxed);
+            let indent = compute_indent(indent_amount);
+
+            // Todo: Recursively ensure that *entire* string is of appropriate
+            // width (not just message).
+            $crate::perf_trace::println!("{}{}", indent, start_msg);
+        }};
+    }
+
     pub fn compute_indent_whitespace(indent_amount: usize) -> String {
         let mut indent = String::new();
         for _ in 0..indent_amount {
@@ -195,7 +216,13 @@ mod inner {
             let _ = $title;
         };
     }
-
+    #[macro_export]
+    macro_rules! add_single_trace {
+        ($msg:expr) => {{
+            let _ = $msg;
+            $crate::perf_trace::TimerInfo
+        }};
+    }
     #[macro_export]
     macro_rules! end_timer {
         ($time:expr, $msg:expr) => {
@@ -219,7 +246,9 @@ mod tests {
 
     #[test]
     fn print_add() {
+        add_single_trace!(|| "Hello");
         let start = start_timer!(|| "Hello");
+        add_single_trace!(|| "HelloWorld");
         add_to_trace!(|| "HelloMsg", || "Hello, I\nAm\nA\nMessage");
         end_timer!(start);
     }